Fixes for self-signed certs: Disable verify and cafile fix.
As referenced in upstream issue #64 (see below), the "cafile"
option seems to have an issue with the Python code for
setting the cafile in the SSL context.

Additionally, the code for "connection_options" was incorrect,
and so you couldn't use it to disable the verification.  This
also has a fix for that and an example of how to use it.

https://github.com/ansible-collections/community.elastic/issues/64